lib.loadOptions now takes an explicit 'commandline' keyword.
The loadOptions function parses either options fed to it or options
obtained from STDIN. It used to choose where to take options based on
whether or not any args were given to it at all. This had the problem
that if no options were given to it, it would pull options from the
calling script whether or not these were intended for it at all. This
can result in silent failure of propka.run.single().

Now, loadOptions takes a keyword argument 'commandline' that, when
False, takes its arguments from its own input. When True, it takes them
from STDIN. The run.single() function also reflect this change.