Molecular_container and run.single() now include a 'writeout' keyword.
This is useful for external codes that don't need the file output, but
would like performance gains from not waiting for a file to be written
to disk each time.

For Molecular_container.__init__(), writeout=False will suppress
writeout of propka inputs. For run.single(), writeout=False will
suppress writeout of both inputs and results.